3D printable V pulley
Two version of an 80t pulley for the V axis, one optimized for printing in resin directly on the build plate, the other optimized for fdm.

I will only be providing .step file going forward since they can be easily edited in CAD software and opened directly in PrusaSlicer for printing or exporting as .stl to open in other applications.